webapi/tct-screenorientation-w3c-tests/screenorientation/blink/lockOrientation-bad-argument.html

<!--
Test convert from Blink:
https://github.com/crosswalk-project/blink-crosswalk/blob/master/LayoutTests/screen_orientation/lockOrientation-bad-argument.html
-->


<!DOCTYPE html>
<meta charset="utf-8">
<title>Screen Test: screen.lockOrientation() returns false on bad input</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://www.w3.org/TR/2012/WD-screen-orientation-20120522/#extensions-to-the-screen-interface" />
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
var t = async_test("Test checks that screen.lockOrientation() returns false on bad input", { timeout: 5000});

function onOrientationChangeEvent(ev) {
  t.step(function () {
    assert_unreached("Unexpected 'orientationchange' event");
  });
  t.done()
}

function completeTest() {
  assert_equals(screen.orientation, "landscape", "default screen orientation");
  screen.unlockOrientation();
}

window.addEventListener("orientationchange", onOrientationChangeEvent);

t.step(function () {
  assert_equals(screen.orientation, "landscape", "default screen orientation");
  assert_false(screen.lockOrientation("invalid-orientation"), "screen.lockOrientation('invalid-orientation')");
  assert_equals(screen.orientation, "landscape", "default screen orientation");
  assert_false(screen.lockOrientation(undefined), "screen.lockOrientation('undefined')");
  assert_equals(screen.orientation, "landscape", "default screen orientation");
  assert_false(screen.lockOrientation(123), "screen.lockOrientation('123')");
  assert_equals(screen.orientation, "landscape", "default screen orientation");
  assert_false(screen.lockOrientation(window), "screen.lockOrientation('window')");
  assert_equals(screen.orientation, "landscape", "default screen orientation");
  assert_false(screen.lockOrientation(['portrait-primary', 'landscape-primary']), "screen.lockOrientation('portrait-primary, landscape-primary')");
  completeTest();
});
t.done()
</script>
